change in structure - loan, loan payment, transaction, account; audit added

+++ src/main/java/rs/ac/bg/fon/ebanking/rest/ClientController.java
@@ -30,7 +30,7 @@ public class ClientController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ClientDTO> findById(@PathVariable("id") Integer id) throws Exception {
+    public ResponseEntity<ClientDTO> findById(@PathVariable("id") Long id) throws Exception {
         return ResponseEntity.ok().body(clientService.findById(id));
     }
